Added a license and README, which shuts github up.
Why am I even putting this on github? Is anyone going to be impressed by this? SSure, it demonstrates that I can read a book and that is a valuable skill to have in today's fast paced information driven society, but its hardly a skill unique to me. In fact literacy rates have been steadily increasing the world over for more than a hundred years (https://ourworldindata.org/literacy), and so while sowing evidence that I can read is certainly better than mere anecdote, its also not enough to significantly boost my standing in the world.

Given how comptetitve the world of tech can seem, it may be that this repository offers little more than cold comfort for my ego.